New Ownership Property: Notes
The owner of a submissive may now "pin a note" to them ("If found,
return to...", "Do not remove $(ITEM_LIST)...," "Is being disciplined
for...," etc.).  The note is free-form (compressed) text stored in the
Ownership struct, and is displayed in the user's OnlineProfile pane.

Create a new Ownership property: `Notes`.  If the updated note has zero
length, the property is set to null.

Create a new Message request type: `ServerAccountOwnershipUpdateNotesRequest`.