Fix owner's notes update back to client
`socket.emit()` is for the account associated with the current
connection/login.  If we want to send Ownership changes to a different
user, we have to send it to their socket.  Hence, when updating owner's
notes, send the Ownership update via `TargetAcc.Socket.emit()`.

Direct the chat room sync toward the `TargetAcc` character as well when
calling `ChatRoomSyncCharacter()`.